Inventory 파일 추가
